build sentinel for master listeners
